# lab4/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PizzGram Api Documentation",
        "description": "L5 Swagger PizzGram Api description",
        "contact": {
            "email": "eliseevv02@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "PizzGram API Server"
        }
    ],
    "paths": {
...
        "/checkout-states": {
            "get": {
                "tags": [
                    "Checkout states"
                ],
                "summary": "Get list of all checkout states",
                "description": "Returns list of checkout states",
                "operationId": "getAllCheckoutStates",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Checkout states"
                ],
                "summary": "Add new checkout state",
                "description": "Adds new checkout state",
                "operationId": "addCheckoutState",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Checkout state name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The adding request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Some checkout state"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The checkout state was created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "checkout-state": {
                                            "$ref": "#/components/schemas/CheckoutState"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/checkout-states/{id}": {
            "get": {
                "tags": [
                    "Checkout states"
                ],
                "summary": "Get one checkout state by id",
                "description": "Returns object of checkout state",
                "operationId": "getCheckoutStateById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checkout state id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "checkout-state": {
                                            "$ref": "#/components/schemas/CheckoutState"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Checkout states"
                ],
                "summary": "Update existing checkout state",
                "description": "Returns updated checkout state data",
                "operationId": "updateCheckoutState",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checkout state id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Checkout state name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The updating request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Some checkout state"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The checkout state was updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CheckoutState"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Checkout states"
                ],
                "summary": "Delete checkout state by id",
                "description": "Deletes checkout state by id",
                "operationId": "deleteCheckoutState",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checkout state id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The checkout state was deleted."
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CheckoutState": {
                "title": "CheckoutState",
                "description": "CheckoutState model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}